+++ bootique-opentelemetry/src/main/java/io/bootique/otel/OpenTelemetryVar.java
@@ -32,8 +32,10 @@ enum OpenTelemetryVar {
     OTEL_SERVICE_NAME("service.name"),
 
     /* Exporter Selection */
-    OTEL_TRACES_EXPORTER("traces.exporter"),
-    OTEL_METRICS_EXPORTER("metrics.exporter"),
+    // TODO: support
+//    OTEL_TRACES_EXPORTER("traces.exporter"),
+//    OTEL_METRICS_EXPORTER("metrics.exporter"),
+//    OTEL_LOGS_EXPORTER("logs.exporter"),
 
     /* OTLP Exporter */
     OTEL_EXPORTER_OTLP_PROTOCOL("exporter.otlp.protocol"),