+++ bootique-opentelemetry/src/main/java/io/bootique/otel/logger/NoneLogsExporterFactory.java
@@ -16,20 +16,18 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package io.bootique.otel.meter;
+package io.bootique.otel.logger;
 
 import com.fasterxml.jackson.annotation.JsonTypeName;
-import io.opentelemetry.exporter.logging.LoggingMetricExporter;
-import io.opentelemetry.sdk.metrics.export.MetricExporter;
 
 /**
  * @since 4.0
  */
-@JsonTypeName("console")
-public class ConsoleMetricExporterFactory implements MetricExporterFactory {
+@JsonTypeName("none")
+public class NoneLogsExporterFactory implements LogsExporterFactory {
 
     @Override
-    public MetricExporter create() {
-        return LoggingMetricExporter.create();
+    public LogRecordExporterHolder create() {
+        return null;
     }
 }