+++ bootique-opentelemetry/src/main/java/io/bootique/otel/meter/SdkMeterProviderFactory.java
@@ -43,7 +43,7 @@ public class SdkMeterProviderFactory {
     private final ShutdownManager shutdownManager;
 
     private Duration metricExportInterval;
-    private List<MetricExporterFactory> metricExporters;
+    private List<MetricsExporterFactory> metricExporters;
 
 
     @Inject
@@ -59,7 +59,7 @@ public class SdkMeterProviderFactory {
     }
 
     @BQConfigProperty
-    public SdkMeterProviderFactory setMetricExporters(List<MetricExporterFactory> metricExporters) {
+    public SdkMeterProviderFactory setMetricExporters(List<MetricsExporterFactory> metricExporters) {
         this.metricExporters = metricExporters;
         return this;
     }
@@ -69,10 +69,9 @@ public class SdkMeterProviderFactory {
                 .builder()
                 .setResource(resource);
 
-        // TODO:
-        //   views
-        //   clock
-        //   exemplar filter
+        // TODO: clock
+        // TODO: views
+        // TODO: exemplar filter
 
         createMetricReaders().forEach(builder::registerMetricReader);
         return shutdownManager.onShutdown(builder.build());
@@ -90,17 +89,18 @@ public class SdkMeterProviderFactory {
         // A single "none" exporter would suppress the default "console" exporter. Though unlike the agent, having a
         // "none" exporter mixed with others doesn't result in an exception. It will just be ignored
 
-        List<MetricExporterFactory> exporters = this.metricExporters == null || this.metricExporters.isEmpty()
-                ? List.of(new ConsoleMetricExporterFactory())
+        List<MetricsExporterFactory> exporters = this.metricExporters == null || this.metricExporters.isEmpty()
+                ? List.of(new ConsoleMetricsExporterFactory())
                 : this.metricExporters;
 
         return exporters.stream()
-                .map(MetricExporterFactory::create)
+                .map(MetricsExporterFactory::create)
                 .filter(Objects::nonNull)
                 .toList();
     }
 
     private MetricReader createMetricReader(MetricExporter exporter) {
+        // No explicit shutdown. The reader is closed by the parent SdkMeterProvider
         return PeriodicMetricReader.builder(exporter)
                 .setInterval(getMetricExportIntervalOrDefault())
                 .build();