+++ RELEASE-NOTES.md
@@ -4,4 +4,5 @@
 * #2 Configurable SdkTracerProvider
 * #3 Injectable OTel Resource with Bootique application name
 * #4 Configurable SdkMeterProvider
+* #5 Configurable SdkLoggerProvider